39 new landing page (#43)
* add Home page for prototype2 and move old routes to /feedback

* add start button to home page

* add live camera feed as background

* update README with HTTPS setup to run client

---------
